Improve space handling to maximise image size
Before, the namebox size has always been substracted. Now, it will use
that space for the image to maximise the size. If a namebox is requested, it will use
available space first (max_height) and if not, it'll shrink the image to fit.
Now, all space is optimally used and the minis are standardized in size.